+++ tests/phaseSpaceGeneration/crossSections/blockB_WW_dilep.lua
@@ -1,7 +1,6 @@
 local p1 = declare_input("p1")
 local p2 = declare_input("p2")
 local p3 = declare_input("p3")
-local p4 = declare_input("p4")
 
 parameters = {
     energy = 13000.,
@@ -11,119 +10,104 @@ parameters = {
 
 cuba = {
     verbosity = 3,
-    max_eval = 1000000000,
+    max_eval = 280000000,
     relative_accuracy = 0.005,
-    n_start = 100000000,
-    n_increase = 100000000,
-    seed = 5468460,
-    ncores = 2,
-    batch_size = 1000000
+    n_start = 10000000,
+    n_increase = 10000000,
+    batch_size = 10000000,
+    ncores = 10,
 }
 
 -- 'Flat' transfer functions to integrate over the visible particle's angles
 
 -- First |P|
+FlatTransferFunctionOnP.tf_p_1 = {
+    ps_point = add_dimension(),
+    reco_particle = p1.reco_p4,
+    min = 0.,
+    max = parameters.energy/2,
+}
+FlatTransferFunctionOnP.tf_p_2 = {
+    ps_point = add_dimension(),
+    reco_particle = p2.reco_p4,
+    min = 0.,
+    max = parameters.energy/2,
+}
 FlatTransferFunctionOnP.tf_p_3 = {
     ps_point = add_dimension(),
     reco_particle = p3.reco_p4,
     min = 0.,
     max = parameters.energy/2,
 }
-p3.set_gen_p4("tf_p_3::output")
-
-FlatTransferFunctionOnP.tf_p_4 = {
-    ps_point = add_dimension(),
-    reco_particle = p4.reco_p4,
-    min = 0.,
-    max = parameters.energy/2,
-}
-p4.set_gen_p4("tf_p_4::output")
 
 -- Then Phi
 FlatTransferFunctionOnPhi.tf_phi_1 = {
     ps_point = add_dimension(),
-    reco_particle = p1.reco_p4,
+    reco_particle = 'tf_p_1::output',
 }
-p1.set_gen_p4("tf_phi_1::output")
-
 FlatTransferFunctionOnPhi.tf_phi_2 = {
     ps_point = add_dimension(),
-    reco_particle = p2.reco_p4,
+    reco_particle = 'tf_p_2::output',
 }
-p2.set_gen_p4("tf_phi_2::output")
-
 FlatTransferFunctionOnPhi.tf_phi_3 = {
     ps_point = add_dimension(),
-    reco_particle = p3.gen_p4,
+    reco_particle = 'tf_p_3::output',
 }
-p3.set_gen_p4("tf_phi_3::output")
-
-FlatTransferFunctionOnPhi.tf_phi_4 = {
-    ps_point = add_dimension(),
-    reco_particle = p4.gen_p4,
-}
-p4.set_gen_p4("tf_phi_4::output")
 
 -- Finally, do Theta
 FlatTransferFunctionOnTheta.tf_theta_1 = {
     ps_point = add_dimension(),
-    reco_particle = p1.gen_p4,
+    reco_particle = 'tf_phi_1::output',
 }
-p1.set_gen_p4('tf_phi_1::output')
-
 FlatTransferFunctionOnTheta.tf_theta_2 = {
     ps_point = add_dimension(),
-    reco_particle = p2.gen_p4,
+    reco_particle = 'tf_phi_2::output',
 }
-p2.set_gen_p4('tf_phi_2::output')
-
 FlatTransferFunctionOnTheta.tf_theta_3 = {
     ps_point = add_dimension(),
-    reco_particle = p3.gen_p4,
+    reco_particle = 'tf_phi_3::output',
 }
-p3.set_gen_p4('tf_phi_3::output')
-
-FlatTransferFunctionOnTheta.tf_theta_4 = {
-    ps_point = add_dimension(),
-    reco_particle = p4.gen_p4,
-}
-p4.set_gen_p4('tf_phi_4::output')
 
 inputs = {
-  p1.gen_p4,
-  p2.gen_p4,
-  p3.gen_p4,
-  p4.gen_p4,
+  'tf_theta_1::output',
+  'tf_theta_2::output',
+  'tf_theta_3::output',
+}
+
+BreitWignerGenerator.flatter_w = {
+    ps_point = add_dimension(),
+    mass = parameter('W_mass'),
+    width = parameter('W_width'),
 }
 
 StandardPhaseSpace.phaseSpaceOut = {
-    particles = { inputs[3], inputs[4] }
+    particles = inputs,
 }
 
-BlockA.blocka = {
-    p1 = inputs[1],
-    p2 = inputs[2],
-    branches = { inputs[3], inputs[4] },
+BlockB.blockb = {
+    s12 = 'flatter_w::s',
+    p2 = inputs[1],
+    branches = { inputs[2], inputs[3] }
 }
 
 -- Loop
 
 Looper.looper = {
-    solutions = "blocka::solutions",
+    solutions = "blockb::solutions",
     path = Path("initial_state", "me_ww", "integrand")
 }
 
-    gen_inputs = { 'looper::particles/1', 'looper::particles/2', inputs[3], inputs[4] }
+    gen_inputs = { 'looper::particles/1', inputs[1], inputs[2], inputs[3] }
 
     BuildInitialState.initial_state = {
         particles = gen_inputs
     }
 
     jacobians = {
-      'tf_p_3::TF_times_jacobian', 'tf_p_4::TF_times_jacobian',
-      'tf_phi_1::TF_times_jacobian', 'tf_phi_2::TF_times_jacobian', 'tf_phi_3::TF_times_jacobian', 'tf_phi_4::TF_times_jacobian',
-      'tf_theta_1::TF_times_jacobian', 'tf_theta_2::TF_times_jacobian', 'tf_theta_3::TF_times_jacobian', 'tf_theta_4::TF_times_jacobian',
-      'phaseSpaceOut::phase_space', 'looper::jacobian',
+      'tf_p_1::TF_times_jacobian', 'tf_p_2::TF_times_jacobian', 'tf_p_3::TF_times_jacobian',
+      'tf_phi_1::TF_times_jacobian', 'tf_phi_2::TF_times_jacobian', 'tf_phi_3::TF_times_jacobian',
+      'tf_theta_1::TF_times_jacobian', 'tf_theta_2::TF_times_jacobian', 'tf_theta_3::TF_times_jacobian',
+      'flatter_w::jacobian', 'phaseSpaceOut::phase_space', 'looper::jacobian',
     }
 
     MatrixElement.me_ww = {